feat: add functions to update quorum pcts (#52)
**Motivation:**

closes #2 

to be closer to Governor, we want to support updating quorums.

**Modifications:**

- Added `setQuorumPct` function
- Added tests
- Added view fns

**Result:**

Quorums can be updated

+++ src/token-voting/LlamaERC20TokenCaster.sol
@@ -34,8 +34,8 @@ contract LlamaERC20TokenCaster is LlamaTokenCaster {
     ILlamaCore _llamaCore,
     ILlamaTokenClockAdapter _clockAdapter,
     uint8 _role,
-    uint256 _voteQuorumPct,
-    uint256 _vetoQuorumPct
+    uint16 _voteQuorumPct,
+    uint16 _vetoQuorumPct
   ) external initializer {
     __initializeLlamaTokenCasterMinimalProxy(_llamaCore, _clockAdapter, _role, _voteQuorumPct, _vetoQuorumPct);
     token = _token;
